+++ cpp/include/cugraph/prims/transform_reduce_v.cuh
@@ -56,7 +56,7 @@ T transform_reduce_v(raft::handle_t const& handle,
                      T init,
                      raft::comms::op_t op = raft::comms::op_t::SUM)
 {
-  auto id = identity<T>(op);
+  auto id = identity_element<T>(op);
   auto ret =
     op_dispatch<T>(op, [&handle, &graph_view, vertex_value_input_first, v_op, id, init](auto op) {
       return thrust::transform_reduce(
