+++ cpp/tests/prims/mg_copy_v_transform_reduce_inout_nbr.cu
@@ -64,8 +64,7 @@ struct property_transform
     operator()(const vertex_t& val)
   {
     cuco::detail::MurmurHash3_32<vertex_t> hash_func{};
-    // auto value = hash_func(val) % mod;
-    auto value = 1;
+    auto value = hash_func(val) % mod;
     return thrust::make_tuple(static_cast<Args>(value)...);
   }
 
@@ -74,8 +73,7 @@ struct property_transform
     const vertex_t& val)
   {
     cuco::detail::MurmurHash3_32<vertex_t> hash_func{};
-    // auto value = hash_func(val) % mod;
-    auto value = 1;
+    auto value = hash_func(val) % mod;
     return static_cast<type>(value);
   }
 };
